% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ocean_acidity_plot_utils_add_hind_data.R
\name{add_hind_data}
\alias{add_hind_data}
\title{add_hind_data
Add some of the hindcast years to the ssps so that the rolling average includes those years}
\usage{
add_hind_data(proj_av, years, is_date = TRUE)
}
\arguments{
\item{proj_av}{Projected model average}

\item{years}{years to add for hindcast}

\item{is_date}{TRUE if there is a date column - will be turned into a year column}
}
\value{
Dataframe of projected model averages with extra hindcast data for the ssps
}
\description{
add_hind_data
Add some of the hindcast years to the ssps so that the rolling average includes those years
}
