% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ocean_acidity_plot_utils_clean_oa_obs.R
\name{clean_oa_obs}
\alias{clean_oa_obs}
\title{clean_oa_obs}
\usage{
clean_oa_obs(which_station, obs_raw)
}
\arguments{
\item{which_station}{String of the station chosen}

\item{obs_raw}{Raw ocean acidity observed data}
}
\value{
Clean dataframe of the ocean acidity observed data
}
\description{
clean_oa_obs
}
