% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_snow_cover_plot_utils_clean_sc_data.R
\name{clean_sc_data}
\alias{clean_sc_data}
\title{clean_sc_data}
\usage{
clean_sc_data(sc_df, min_hind_yr, years)
}
\arguments{
\item{sc_df}{Dataframe with combined projected and observed snow cover data}

\item{min_hind_yr}{Minimum year we want the hindcast data to go to}

\item{years}{Years to feed into add_hind_data function}
}
\value{
Dataframe of tidyed projected average and observed data and a new rolling average column
}
\description{
clean_sc_data
}
\examples{
clean_sc_data(sc_df, 1950, c(2009:2014))
}
