% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ghg_conc_plot_utils_cln_ghg_obs.R
\name{cln_ghg_obs}
\alias{cln_ghg_obs}
\title{cln_ghg_obs}
\usage{
cln_ghg_obs(raw_obs, rename_yr = FALSE)
}
\arguments{
\item{raw_obs}{Dataframe of raw observed ghg concentration}

\item{rename_yr}{True if the year column is called `Year (negative values = BC)`}
}
\value{
Clean dataframe of observed ghg data
}
\description{
cln_ghg_obs
}
