% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ghg_conc_plot_utils_create_ghg_plot.R
\name{create_ghg_plot}
\alias{create_ghg_plot}
\title{create_ghg_plot}
\usage{
create_ghg_plot(observed_data, projected_data, which_ghg, unit, which_colors)
}
\arguments{
\item{observed_data}{Dataframe of observed ghg data filtered to which ghg type to plot}

\item{projected_data}{Dataframe of projected ghg data filtered to which ghg type to plot}

\item{which_ghg}{String of which ghg type to plot}

\item{unit}{String of the unit of ghg type}

\item{which_colors}{List of the colors to use for plotting the projected data}
}
\value{
Highchart plot
}
\description{
create_ghg_plot
}
