% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_snow_cover_plot_utils_create_sc_seas.R
\name{create_sc_seas}
\alias{create_sc_seas}
\title{create_sc_seas}
\usage{
create_sc_seas(seas_num, seas_name, min_hind_yr)
}
\arguments{
\item{seas_num}{The number for the season}

\item{seas_name}{String of the name of the season}

\item{min_hind_yr}{The minimum hindcast year}
}
\value{
Snow cover dataframe ready for plotting
}
\description{
create_sc_seas
}
\examples{
create_sc_seas(seas_num = 2, seas_name = "Spring", min_hind_yr = 1950)
}
