% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_slr_map_utils_create_slr_station_plot.R
\name{create_slr_station_plot}
\alias{create_slr_station_plot}
\title{create_slr_station_plot}
\usage{
create_slr_station_plot(slr_stations_all, which_station)
}
\arguments{
\item{slr_stations_all}{df of all the slr stations and scenarios}

\item{which_station}{string of which station to plot}
}
\description{
create_slr_station_plot
}
