% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_generate_static_maps.R
\name{generate_static_maps}
\alias{generate_static_maps}
\title{Generate and save static maps for each scenario (if not already saved)}
\usage{
generate_static_maps(
  cln_data,
  module_id,
  create_static_map_fn = create_static_map,
  width = 1000,
  height = 600,
  ...
)
}
\arguments{
\item{cln_data}{An `sf` object with a `scenario` column}

\item{module_id}{String, name of the module (used in file path)}

\item{create_static_map_fn}{Function to generate the map}

\item{width}{Width of image in pixels (default 1000)}

\item{height}{Height of image in pixels (default 600)}

\item{...}{Additional arguments passed to `create_static_map_fn`}
}
\value{
Named list mapping scenario name → relative image path
}
\description{
Generate and save static maps for each scenario (if not already saved)
}
