% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/mod_ocean_acidity_plot_utils_process_oa_station.R
\name{process_oa_station}
\alias{process_oa_station}
\title{process_oa_station}
\usage{
process_oa_station(which_station, obs_data, mod_av, mod_all)
}
\arguments{
\item{which_station}{String of which ocean acidity station}

\item{obs_data}{Dataframe of observed data}

\item{mod_av}{Dataframe of model average data}

\item{mod_all}{Dataframe of all modeled data}
}
\description{
process_oa_station
}
