% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_snow_cover_plot_utils_process_sc.R
\name{process_sc}
\alias{process_sc}
\title{process_sc}
\usage{
process_sc(model_av_raw, all_model_raw, min_hind_yr, which_season)
}
\arguments{
\item{model_av_raw}{Dataframe of projected and observed data}

\item{all_model_raw}{Dataframe with all of the models}

\item{min_hind_yr}{Minimum year to filter the hindcast data to}

\item{which_season}{String of the season to filter the data to}
}
\description{
process_sc
}
\examples{
ann_sc_all <- process_sc(sc_an_proj_av, sc_an_proj_mod, min_hind_yr)
}
