% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_arctic_sea_ice_plot_utils_process_sea_ice.R
\name{process_sea_ice}
\alias{process_sea_ice}
\title{process_sea_ice}
\usage{
process_sea_ice(obs_proj_av, ssp_data, start_year, end_year, min_hind_yr)
}
\arguments{
\item{obs_proj_av}{Dataframe with observed and projected average}

\item{ssp_data}{Dataframe with all the projected models}

\item{start_year}{Year the overlap between the hindcast and observed data started}

\item{end_year}{Year the overlap between the hindcast and observed data ended}

\item{min_hind_yr}{The minimum year we want the hindcst}
}
\value{
Dataframe ready for plotting
}
\description{
process_sea_ice
}
\examples{
asi_adj_all <- process_sea_ice(asi_proj_av_all, asi_proj_mod, base_yr_start, base_yr_end, min_hind_yr)
}
