% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_seasonal_temp_plot_utils_process_seasons.R
\name{process_seasons}
\alias{process_seasons}
\title{process_seasons}
\usage{
process_seasons(
  which_season,
  obs_data,
  proj_data,
  ssp_data,
  ssp_var,
  base_yr_start,
  base_yr_end
)
}
\arguments{
\item{which_season}{Which season to filter to (e.g. "Fall", "Spring". etc.)}

\item{obs_data}{Dataframe of observed data}

\item{proj_data}{Dataframe of projected average data}

\item{ssp_data}{Dataframe of all projected models}

\item{ssp_var}{Column name of the temperature variable for the all project models dataframe}

\item{base_yr_start}{Baseline start year}

\item{base_yr_end}{Baseline end year}
}
\value{
Adjusted model projections and observed dataframe for selected season
}
\description{
process_seasons
}
\examples{
seas_proj_adj <- process_seasons(which_season = "Fall",obs_data = obs_raw,proj_data = proj_av,ssp_data = proj_all,ssp_var = avg_temp_f,base_yr_start = base_yr_start,base_yr_end = base_yr_end)
}
