% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_ocean_acidity_plot_utils_rename_stations.R
\name{rename_stations}
\alias{rename_stations}
\title{rename_stations}
\usage{
rename_stations(proj_av)
}
\arguments{
\item{proj_av}{The projected model average dataframe for ocean acidity}
}
\value{
Dataframe with new column that gives the full station name for ocean acidity
}
\description{
rename_stations
}
