% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_slr_plot_utils_create_slr_plot.R
\name{create_slr_plot}
\alias{create_slr_plot}
\title{create_slr_plot}
\usage{
create_slr_plot(slr_obs_cln, slr_proj_all, csiro_bounds)
}
\arguments{
\item{slr_obs_cln}{Dataframe of observed slr data}

\item{slr_proj_all}{Dataframe of projected slr data}

\item{csiro_bounds}{Dataframe of CSIRO bounds for the observed slr data}
}
\description{
create_slr_plot
}
